% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_wormhole.R
\name{add_wormhole}
\alias{add_wormhole}
\title{Add a Manual Wormhole Edge}
\usage{
add_wormhole(nb, wormhole, coord)
}
\arguments{
\item{nb}{An nb neighbor list object.}

\item{wormhole}{A 1-row data frame/tibble containing projected metrics: `from_x`, `from_y`, `to_x`, `to_y`.}

\item{coord}{A 2-column matrix of projected grid coordinates (UTM).}
}
\value{
An updated 'nb' object with the added edge.
}
\description{
Maps single path anchors into logical neighborhood paths based on closest centroids.
}
