% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nb_set_edge}
\alias{nb_set_edge}
\title{Symmetrically Set a Graph Edge}
\usage{
nb_set_edge(nb, i, j)
}
\arguments{
\item{nb}{An 'nb' neighbor list object.}

\item{i}{Integer index of first node.}

\item{j}{Integer index of second node.}
}
\description{
Adds undirected links to an nb list object.
}
\keyword{internal}
