% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stitch_island.R
\name{stitch_island}
\alias{stitch_island}
\title{Stitch Isolated Islands}
\usage{
stitch_island(nb, coord)
}
\arguments{
\item{nb}{An 'nb' neighbor list object.}

\item{coord}{A 2-column matrix of projected grid coordinates (UTM).}
}
\value{
An updated 'nb' object.
}
\description{
Explicitly isolates regions with zero neighbors and forces a topological connection to their nearest mainland counterpart.
}
