% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarry_stitch_nb.R
\name{summary.stitch_nb}
\alias{summary.stitch_nb}
\title{Summary of a Stitched Neighborhood Graph}
\usage{
\method{summary}{stitch_nb}(object, ...)
}
\arguments{
\item{object}{An object of class `stitch_nb`.}

\item{...}{Additional arguments affecting the summary produced.}
}
\description{
Summary of a Stitched Neighborhood Graph
}
