% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweep_absorp.R
\name{sweep_absorp}
\alias{sweep_absorp}
\title{Sweep Absorption Pass}
\usage{
sweep_absorp(nb, coord, max_pass = 10)
}
\arguments{
\item{nb}{An nb neighbor list object.}

\item{coord}{A 2-column matrix of projected grid coordinates (UTM).}

\item{max_pass}{Max allowed graph evaluation loops. Defaults to 10.}
}
\value{
Updated 'nb' object.
}
\description{
Iteratively evaluates disjoint sub-graph fragments, snapping them to the main giant network component.
}
